requirements.txt
# WriteLens - Root Requirements
# Consolidated dependencies from all Python modules

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Machine Learning & Statistics
scikit-learn>=1.3.0
pgmpy>=0.1.25
simple-statistics>=1.0.0

# NLP & Text Analysis
nltk>=3.8.0
textstat>=0.7.2
spacy>=3.7.0

# Utilities
openpyxl>=3.1.0
xlrd>=2.0.0
